Adds input capabilities.
The arena is now able to handle input. It is useful to speed up the
processing time during any match.

The Arena class is responsible for dealing with some special inputs,
but the rendering classes should take care of listening to input
sources. I did that because in the future a special render method may
use different input sources, such as a specific class from any game
engine.